Update readme and refactor to hand.py
